+++ app/[lang]/(authorization)/blog/page.jsx
