+++ app/[lang]/(authorization)/page.js
@@ -1,4 +1,4 @@
-import { MainContent } from "../../components/mainContent/MainContent";
+import { MainContent } from "../components/mainContent/MainContent";
 
 export default function Home() {
   return (
