+++ app/[lang]/(authorization)/product/[id]/page.jsx
