+++ app/[lang]/(authorization)/singleBlog/[id]/page.jsx
