+++ app/[lang]/components/Loading.jsx
