+++ app/[lang]/components/blog/Blog.jsx
