+++ app/[lang]/components/headerFooterContent/Header.jsx
@@ -1,13 +1,13 @@
 "use server";
 import Image from "next/image";
 // import { GiHamburgerMenu } from "react-icons/gi";
-import { Navigation } from "../headerFooterContent/Navigation";
-import logo from "../../public/assets/image/footerLogo.webp";
+import { Navigation } from "./Navigation";
+import logo from "../../../../public/assets/image/footerLogo.webp";
 // import { MobileMenu } from "../headerFooterContent/MobileMenu";
 // import { useState } from "react";
-import { HeaderIcon } from "../headerFooterContent/HeaderIcon";
+import { HeaderIcon } from "./HeaderIcon";
 import Link from "next/link";
-import { logout } from "../../app/actions";
+// import { logout } from "../../app/actions";
 import { LogOutButton } from "../logOutButton/LogOutButton";
 import ThemeMode from "./ThemeMode";
 
@@ -15,10 +15,10 @@ async function Header() {
   // const [open, setOpen] = useState(false);
 
   // function for user log out
-  const handleLogOut = async () => {
-    "use server";
-    await logout();
-  };
+  // const handleLogOut = async () => {
+  //   "use server";
+  //   await logout();
+  // };
 
   return (
     <header className="w-full bg-[#111827] py-4 px-0 border-b border-[#939898] dark:bg-yellow-400">
@@ -36,7 +36,7 @@ async function Header() {
           <Navigation marginLeft="30px" display="flex" />
         </div>
         {/* <HeaderIcon marginRight="20px" alignItmes="center" /> */}
-        <LogOutButton handleLogOut={handleLogOut} />
+        <LogOutButton />
         <ThemeMode />
       </div>
 
@@ -46,3 +46,4 @@ async function Header() {
 }
 
 export default Header;
+//handleLogOut = { handleLogOut };
