+++ app/[lang]/components/headerFooterContent/HeaderIcon.jsx
