+++ app/[lang]/components/headerFooterContent/MobileMenu.jsx
