+++ app/[lang]/components/headerFooterContent/SocialLinks.jsx
