+++ app/[lang]/components/mainContent/MainContent.jsx
