+++ app/[lang]/logIn/page.js
@@ -1,19 +1,19 @@
 "use server";
-import { cookies } from "next/headers";
-import { AUTH_COOKIE_KEY } from "../contants";
-import { redirect } from "next/navigation";
-import { UserAuthorization } from "../../components/userLogIn/UserAuthorization";
-import { Userlogin } from "../actions";
+// import { cookies } from "next/headers";
+// import { AUTH_COOKIE_KEY } from "../contants";
+// import { redirect } from "next/navigation";
+import { UserAuthorization } from "../components/userLogIn/UserAuthorization";
+import { Userlogin } from "../../actions";
 
 export async function LogIn() {
-  const cookieStore = cookies();
-  const cookie = cookieStore.get(AUTH_COOKIE_KEY);
+  // const cookieStore = cookies();
+  // const cookie = cookieStore.get(AUTH_COOKIE_KEY);
 
-  if (cookie?.value) {
-    redirect("/");
-  }
+  // if (cookie?.value) {
+  //   redirect("/");
+  // }
 
-  // function to set user info
+  //function to set user info
   const handleLogIn = async (username, password) => {
     "use server";
     await Userlogin(username, password);
